# civic-data-trust/openapi.json
{"openapi":"3.1.0","info":{"title":"Civic Data Trust Backend","description":"Civic Data Trust Backend API - A comprehensive platform for managing civic data and user roles","version":"1.0.0"},"paths":{"/api/v1/users/":{"get":{"tags":["users"],"summary":"Get Users","operationId":"get_users_api_v1_users__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/app__schemas__user__UserResponse"},"title":"Response Get Users Api V1 Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/{user_id}":{"get":{"tags":["users"],"summary":"Get User","operationId":"get_user_api_v1_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__schemas__user__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["users"],"summary":"Update User","operationId":"update_user_api_v1_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__schemas__user__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["users"],"summary":"Delete User","operationId":"delete_user_api_v1_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/add-to-community/{community_id}":{"post":{"tags":["users"],"summary":"Add User To Community","operationId":"add_user_to_community_api_v1_users_add_to_community__community_id__post","parameters":[{"name":"community_id","in":"path","required":true,"schema":{"type":"string","title":"Community Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddUserToCommunityResponse"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddUserToCommunityResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-roles/":{"get":{"tags":["user-roles"],"summary":"Get All User Roles","operationId":"get_all_user_roles_api_v1_user_roles__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserRoleResponse"},"type":"array","title":"Response Get All User Roles Api V1 User Roles  Get"}}}}}},"post":{"tags":["user-roles"],"summary":"Create User Role","operationId":"create_user_role_api_v1_user_roles__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-roles/user/{user_id}":{"get":{"tags":["user-roles"],"summary":"Get User Roles By User","operationId":"get_user_roles_by_user_api_v1_user_roles_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRoleResponse"},"title":"Response Get User Roles By User Api V1 User Roles User  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-roles/role/{role_id}":{"get":{"tags":["user-roles"],"summary":"Get User Roles By Role","operationId":"get_user_roles_by_role_api_v1_user_roles_role__role_id__get","parameters":[{"name":"role_id","in":"path","required":true,"schema":{"type":"string","title":"Role Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRoleResponse"},"title":"Response Get User Roles By Role Api V1 User Roles Role  Role Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-roles/{user_role_id}":{"delete":{"tags":["user-roles"],"summary":"Delete User Role","operationId":"delete_user_role_api_v1_user_roles__user_role_id__delete","parameters":[{"name":"user_role_id","in":"path","required":true,"schema":{"type":"string","title":"User Role Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["user-roles"],"summary":"Update User Role","operationId":"update_user_role_api_v1_user_roles__user_role_id__put","parameters":[{"name":"user_role_id","in":"path","required":true,"schema":{"type":"string","title":"User Role Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRoleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user-roles/user/{user_id}/role/{role_id}":{"delete":{"tags":["user-roles"],"summary":"Delete User Role By User And Role","operationId":"delete_user_role_by_user_and_role_api_v1_user_roles_user__user_id__role__role_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"role_id","in":"path","required":true,"schema":{"type":"string","title":"Role Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["authentication"],"summary":"Signup","operationId":"signup_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignup"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__schemas__auth__UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["authentication"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["authentication"],"summary":"Refresh Token","operationId":"refresh_token_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["authentication"],"summary":"Logout","operationId":"logout_api_v1_auth_logout_post","responses":{"204":{"description":"Successful Response"}}}},"/api/v1/auth/password-reset-request":{"post":{"tags":["authentication"],"summary":"Password Reset Request","operationId":"password_reset_request_api_v1_auth_password_reset_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/password-reset":{"post":{"tags":["authentication"],"summary":"Password Reset","operationId":"password_reset_api_v1_auth_password_reset_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordReset"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/role-base-modules/":{"post":{"tags":["role-base-modules"],"summary":"Create Role Base Module","operationId":"create_role_base_module_api_v1_role_base_modules__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleBaseModuleCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["role-base-modules"],"summary":"List Role Base Modules","operationId":"list_role_base_modules_api_v1_role_base_modules__get","parameters":[{"name":"module_name","in":"query","required":false,"schema":{"type":"string","description":"Filter by module name (partial match)","title":"Module Name"},"description":"Filter by module name (partial match)"},{"name":"user_role_id","in":"query","required":false,"schema":{"type":"string","description":"Filter by user role ID","title":"User Role Id"},"description":"Filter by user role ID"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number for pagination","default":1,"title":"Page"},"description":"Page number for pagination"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Number of items per page","default":10,"title":"Limit"},"description":"Number of items per page"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/role-base-modules/{module_id}":{"get":{"tags":["role-base-modules"],"summary":"Get Role Base Module By Id","operationId":"get_role_base_module_by_id_api_v1_role_base_modules__module_id__get","parameters":[{"name":"module_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the role base module","title":"Module Id"},"description":"ID of the role base module"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["role-base-modules"],"summary":"Update Role Base Module","operationId":"update_role_base_module_api_v1_role_base_modules__module_id__put","parameters":[{"name":"module_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the role base module to update","title":"Module Id"},"description":"ID of the role base module to update"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleBaseModuleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/role-base-modules/{module_id}/add-roles":{"put":{"tags":["role-base-modules"],"summary":"Add User Roles To Module","operationId":"add_user_roles_to_module_api_v1_role_base_modules__module_id__add_roles_put","parameters":[{"name":"module_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the role base module to update","title":"Module Id"},"description":"ID of the role base module to update"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleBaseModuleUserRoleOperation"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/role-base-modules/{module_id}/remove-roles":{"put":{"tags":["role-base-modules"],"summary":"Remove User Roles From Module","operationId":"remove_user_roles_from_module_api_v1_role_base_modules__module_id__remove_roles_put","parameters":[{"name":"module_id","in":"path","required":true,"schema":{"type":"string","description":"ID of the role base module to update","title":"Module Id"},"description":"ID of the role base module to update"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleBaseModuleUserRoleOperation"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/":{"get":{"tags":["community"],"summary":"Get Communities","description":"Get all communities","operationId":"get_communities_api_v1_community__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":10,"title":"Limit"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/{community_id}":{"get":{"tags":["community"],"summary":"Get Community","description":"Get a community by community_id","operationId":"get_community_api_v1_community__community_id__get","parameters":[{"name":"community_id","in":"path","required":true,"schema":{"type":"string","title":"Community Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/create":{"post":{"tags":["community"],"summary":"Create Community","description":"Create a new community with the provided data.","operationId":"create_community_api_v1_community_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityCreate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/update/{community_id}":{"put":{"tags":["community"],"summary":"Update Community","description":"Update an existing community by community_id.","operationId":"update_community_api_v1_community_update__community_id__put","parameters":[{"name":"community_id","in":"path","required":true,"schema":{"type":"string","title":"Community Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/delete/{community_id}":{"delete":{"tags":["community"],"summary":"Delete Community","description":"Delete a community by community_id.","operationId":"delete_community_api_v1_community_delete__community_id__delete","parameters":[{"name":"community_id","in":"path","required":true,"schema":{"type":"string","title":"Community Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/add-user":{"put":{"tags":["community"],"summary":"Add User To Community","description":"Add a user to a community using community_id and user_id.","operationId":"add_user_to_community_api_v1_community_add_user_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityAddUserRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/remove-user":{"put":{"tags":["community"],"summary":"Remove User From Community","description":"Remove a user from a community using community_id and user_id.","operationId":"remove_user_from_community_api_v1_community_remove_user_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityAddUserRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/request-join":{"post":{"tags":["community"],"summary":"Request To Join Community","description":"Request to join a community using community_id and user_id.","operationId":"request_to_join_community_api_v1_community_request_join_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityJoinRequestCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/confirm-join-request":{"post":{"tags":["community"],"summary":"Confirm Join Request","description":"Confirm or reject a user's request to join a community using request_id.","operationId":"confirm_join_request_api_v1_community_confirm_join_request_post","parameters":[{"name":"request_id","in":"query","required":true,"schema":{"type":"string","title":"Request Id"}},{"name":"approve","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Approve"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/community/join-requests/{community_id}":{"get":{"tags":["community"],"summary":"Get Join Requests","description":"Get all join requests for a community using community_id.","operationId":"get_join_requests_api_v1_community_join_requests__community_id__get","parameters":[{"name":"community_id","in":"path","required":true,"schema":{"type":"string","title":"Community Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/datasets/create":{"post":{"tags":["datasets"],"summary":"Create Shared File","operationId":"create_shared_file_api_v1_datasets_create_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_shared_file_api_v1_datasets_create_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Read Root","description":"Root endpoint providing API information and available endpoints.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Health check endpoint to verify API status and database connectivity.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AddUserToCommunityResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"community_id":{"type":"string","title":"Community Id"},"added_at":{"type":"string","format":"date-time","title":"Added At"}},"type":"object","required":["user_id","community_id","added_at"],"title":"AddUserToCommunityResponse"},"AuthInfo":{"properties":{"is_active":{"type":"boolean","title":"Is Active"},"email_verified":{"type":"boolean","title":"Email Verified"},"last_login":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login"},"failed_login_attempts":{"type":"integer","title":"Failed Login Attempts","default":0},"locked_until":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Locked Until"}},"type":"object","required":["is_active","email_verified","last_login","failed_login_attempts","locked_until"],"title":"AuthInfo","description":"Schema for authentication information."},"Body_create_shared_file_api_v1_datasets_create_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_create_shared_file_api_v1_datasets_create_post"},"CommunityAddUserRequest":{"properties":{"community_id":{"type":"string","title":"Community Id"},"user_id":{"type":"string","title":"User Id"}},"type":"object","required":["community_id","user_id"],"title":"CommunityAddUserRequest"},"CommunityCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"CommunityCreate"},"CommunityJoinRequestCreate":{"properties":{"community_id":{"type":"string","title":"Community Id"},"user_id":{"type":"string","title":"User Id"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["community_id","user_id"],"title":"CommunityJoinRequestCreate"},"CommunityListResponse":{"properties":{"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"},"data":{"items":{"$ref":"#/components/schemas/CommunityResponse"},"type":"array","title":"Data"}},"type":"object","required":["status","message","data"],"title":"CommunityListResponse"},"CommunityResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"logo":{"type":"string","title":"Logo"}},"type":"object","required":["id","name","logo"],"title":"CommunityResponse"},"CommunityUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","title":"CommunityUpdate"},"DatasetResponse":{"properties":{"id":{"type":"string","title":"Id"}},"type":"object","required":["id"],"title":"DatasetResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"user":{"$ref":"#/components/schemas/app__schemas__auth__UserResponse"}},"type":"object","required":["access_token","refresh_token","user"],"title":"LoginResponse"},"PasswordReset":{"properties":{"token":{"type":"string","title":"Token"},"new_password":{"type":"string","minLength":8,"title":"New Password","description":"New password must be at least 8 characters"}},"type":"object","required":["token","new_password"],"title":"PasswordReset","description":"Schema for password reset."},"PasswordResetRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"PasswordResetRequest","description":"Schema for password reset request."},"RefreshTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshTokenRequest","description":"Schema for refresh token request."},"RoleBaseModuleCreate":{"properties":{"module_name":{"type":"string","maxLength":100,"minLength":2,"title":"Module Name","description":"Name of the module"},"user_roles":{"items":{"type":"string"},"type":"array","title":"User Roles","description":"List of user role IDs associated with this module"}},"type":"object","required":["module_name"],"title":"RoleBaseModuleCreate","description":"Schema for creating a new role base module."},"RoleBaseModuleUpdate":{"properties":{"module_name":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Module Name","description":"Updated name of the module"},"user_roles":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"User Roles","description":"Updated list of user role IDs associated with this module"}},"type":"object","title":"RoleBaseModuleUpdate","description":"Schema for updating a role base module."},"RoleBaseModuleUserRoleOperation":{"properties":{"user_role_ids":{"items":{"type":"string"},"type":"array","minItems":1,"title":"User Role Ids","description":"List of user role IDs to add or remove"}},"type":"object","required":["user_role_ids"],"title":"RoleBaseModuleUserRoleOperation","description":"Schema for adding or removing user roles from a module."},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token","token_type"],"title":"Token","description":"Schema for authentication tokens."},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password","description":"Password must be at least 8 characters"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"Schema for user login."},"UserRoleCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"UserRoleCreate"},"UserRoleResponse":{"properties":{"name":{"type":"string","title":"Name"},"id":{"type":"string","title":"Id"}},"type":"object","required":["name","id"],"title":"UserRoleResponse"},"UserSignup":{"properties":{"name":{"type":"string","maxLength":100,"minLength":2,"title":"Name","description":"User's full name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password","description":"Password must be at least 8 characters"},"role":{"type":"string","title":"Role","description":"List of role IDs to assign to the user"}},"type":"object","required":["name","email","password","role"],"title":"UserSignup","description":"Schema for user registration."},"UserUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"status":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Status"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"app__schemas__auth__UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"status":{"type":"boolean","title":"Status"},"role":{"type":"string","title":"Role"},"auth":{"anyOf":[{"$ref":"#/components/schemas/AuthInfo"},{"type":"null"}]},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","email","status","role","auth","created_at","updated_at"],"title":"UserResponse","description":"Schema for user data response (without sensitive information)."},"app__schemas__user__UserResponse":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"status":{"type":"boolean","title":"Status","default":true},"role":{"type":"string","title":"Role"},"id":{"type":"string","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","email","status","role","id","created_at","updated_at"],"title":"UserResponse"}}}}